refactored parser file to contain only globally needed ressources inside module
